fix delete_files to allow work on restore_files

+++ recycle_bin.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
-set -e
+#set -e
 
 #################################################
 # Linux Recycle Bin Simulation
@@ -10,8 +10,9 @@ set -e
 #################################################
 
 # Global Variables (ALL CAPS)
-RECYCLE_BIN_DIR="./recycle_bin"
+RECYCLE_BIN_DIR="$HOME/recycle_bin"
 METADATA_FILE="$RECYCLE_BIN_DIR/metadata.db"
+
 
 #################################################
 # Function: initialize_recyclebin
@@ -30,6 +31,7 @@ initialize_recyclebin(){
         printf "MAX_SIZE_MB=1024\n" > "$RECYCLE_BIN_DIR/config"
     fi 
     touch "$RECYCLE_BIN_DIR/recyclebin.log"
+    echo "Recycle Bin initialized sucessfully"
     return 0
 }
 
@@ -71,7 +73,7 @@ generate_unique_id() {
 get_metadata(){
     if [ $# -ne 1 ]; then
         echo "Erro: get_metadata takes 1 argument only"
-        echo "$(date +"[%d/%m/%Y %H:%M:%S]"): ERROR: get_metadata takes 1 argument only" >> $RECYCLE_BIN_DIR/recyclebin.log
+        log "ERROR: get_metadata takes 1 argument only"
         return 1
     fi
     local file="$1"
@@ -103,7 +105,7 @@ get_metadata(){
 # Description: Move files/directories to recycle bin
 # Parameters: At least 1 (file/directory paths)
 # Returns: 0 on success
-#################################################
+#################################################           TODO:recursividade
 delete_file(){
     #ficheiros protegidos (não podem ser apagados)
     local protected_files=(
@@ -112,16 +114,17 @@ delete_file(){
     )
 
     #Para cada um dos argumentos fazer o seguinte
-    for var in "$@"; do
+    for varr in "$@"; do
+        
+        local var="$varr"
+
         #Skip invalid input
         if [ ! -e "$var" ]; then
             echo "\"$var\" isn't a filename or directory"
             log "Failed to delete $(pwd)/$var - There is no such File/Dir"
             continue
         fi
-
         local abs_path=$(realpath "$var" 2>/dev/null)
-
         # Proteção contra auto-deleção do recycle bin
         if [[ "$abs_path" == "$(realpath "$RECYCLE_BIN_DIR" 2>/dev/null)"* ]] ||
            [[ "$abs_path" == "$HOME/.recycle_bin"* ]]; then
@@ -129,7 +132,6 @@ delete_file(){
             log "Attempted to delete recycle bin: $abs_path"
             continue
         fi
-
         # Verificar ficheiros protegidos
         local is_protected=0
         for protected in "${protected_files[@]}"; do
@@ -141,17 +143,29 @@ delete_file(){
                 break
             fi
         done
-
         if [ $is_protected -eq 1 ]; then
             continue
         fi
 
-        local current_id=$(generate_unique_id)
-        echo "$current_id,$(get_metadata $var)" >> $METADATA_FILE
-        mv "$var" "$RECYCLE_BIN_DIR/files/$current_id"
+        if [[ -f $var ]];then
+            local current_id=$(generate_unique_id)
+            echo "$current_id,$(get_metadata $var)" >> $METADATA_FILE
+            mv "$var" "$RECYCLE_BIN_DIR/files/$current_id"
+            echo "$(realpath "$var") was deleted"
+            log "$(realpath "$var") Was deleted; ID:$current_id"
+        elif [[ -d $var ]];then
 
-        echo "$(realpath "$var") was deleted"
-        log "$(realpath "$var") Was deleted; ID:$current_id"
+            for recursive_var in "$var"/*;do
+                delete_file "$recursive_var"
+            done
+
+            local current_id=$(generate_unique_id)
+            echo "$current_id,$(get_metadata $var)" >> $METADATA_FILE
+            mv "$var" "$RECYCLE_BIN_DIR/files/$current_id"
+            echo "$(realpath "$var") was deleted"
+            log "$(realpath "$var") Was deleted; ID:$current_id"
+        fi
+
     done
     return 0
 }
@@ -161,8 +175,22 @@ delete_file(){
 # Description: List all items in recycle bin
 # Parameters: 0 or 1 (--detailed flag)
 # Returns: 0 on success
-#################################################
+#################################################               TODO: File size readable format, empty filebin
 list_recycled(){
+
+    echo
+    local total_item=$(ls "$RECYCLE_BIN_DIR/files" -1 | wc -l)
+    local total_storage=$(
+        local total=0
+        while IFS= read -r line; do
+
+            IFS=',' read -ra fields <<< "$line"
+            total=$((total + ${fields[4]}))
+
+        done < <(tail -n +2 "$METADATA_FILE")
+        echo $total
+    )
+
     if [ "$#" -eq 0 ]; then
         awk -F',' '{print "| " $1 , "+| "$2 , "+| " $4 , "+| " $5 , "+|"}' $METADATA_FILE | column -t -s+
     elif [ "$#" -eq 1 ] && [ "$1" == "--detailed" ]; then
@@ -172,7 +200,13 @@ list_recycled(){
     else
         echo "list_recycled can only take one argument"
     fi
+
+    printf "\nTotal files: $total_item\n"
+    printf "Total files: $total_storage\n"
+
     echo
+
+
 }
 
 #################################################
@@ -203,9 +237,6 @@ empty_recyclebin(){
                 # Reset do metadata file (mantém apenas o header)
                 head -n 1 "$METADATA_FILE" > "$METADATA_FILE.tmp"
                 mv "$METADATA_FILE.tmp" "$METADATA_FILE"
-
-                # Reset do ID counter no config
-                sed -i '2c0' "$RECYCLE_BIN_DIR/config"
 
                 echo "Recycle bin emptied successfully"
                 log "Recycle bin emptied - all items permanently deleted"
@@ -262,6 +293,76 @@ empty_recyclebin(){
     return 0
 }
 
+#################################################
+# Function: restore_files
+# Description: Restore files by id
+# Parameters: At least 1 (ID)
+# Returns: 0 on success
+#################################################
+restore_files(){
+
+    if [[ "$#" -lt 1 ]]; then
+        echo "Restore Files needs ate least 1 argument"
+        return 1
+    fi
+
+    for var in "$@"; do
+
+        #Modo ID
+        if [[ $var =~ ^[0-9]{10}_[a-zA-Z0-9]{6}$ ]]; then
+
+            # Verificar se o ID existe no metadata
+            if ! grep -q "^$var," "$METADATA_FILE"; then
+                echo "Error: ID '$var' not found in recycle bin"
+                log "Failed to delete item with ID '$var' - ID not found"
+                return 1
+            fi
+
+            #Split metadata into array
+            IFS=',' read -ra arr <<< "$(grep "^$var," "$METADATA_FILE")"
+
+            #check if file is already there
+            if [[ -e "${arr[2]}" ]];then
+
+                # Confirmar com o utilizador
+                echo "WARNING: File already Exists"
+                echo "Do you wish to Override (O), Restore with modified name (M), or Cancel (c) "
+                read -r confirmation
+
+                case "$confirmation" in
+                    O)
+                        mv $RECYCLE_BIN_DIR/files/$var "${arr[2]}"
+                        chmod ${arr[6]} "${arr[2]}"
+                        ;;
+                    M)
+
+                        ;;
+                    *)
+                        echo "Operation cancelled"
+                        return 0
+                        ;;
+                esac
+
+            else
+                mkdir -p "${arr[2]}"
+                mv "$RECYCLE_BIN_DIR/files/$var" "${arr[2]}"
+                chmod ${arr[6]} "${arr[2]}"
+
+            fi
+
+                grep -v "^$var," "$METADATA_FILE" > "$METADATA_FILE.tmp"
+                mv "$METADATA_FILE.tmp" "$METADATA_FILE"  
+
+                echo "File ID $var was restored sucessfully to ${arr[2]}"
+                log "$var restored to ${arr[2]}"
+
+        fi
+
+    done
+    return 0
+
+}
+
 
 #################################################
 # Function: main
@@ -271,7 +372,7 @@ empty_recyclebin(){
 #################################################
 main(){
 
-    first_arg=$1
+    local first_arg=$1
     shift
 
     case $first_arg in
@@ -292,12 +393,16 @@ main(){
         empty_recyclebin "$@"
         ;;
 
+        restore_files | 4)
+        restore_files "$@"
+        ;;
+
         *)
         echo "Unknown command"
         ;;
 
     esac
-
+    return 0
 }
 
 main $@
